feat: refactor filesystem routing into files and routes. Use Wax instead of glob
