+++ src/router/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod files;
+pub mod route;
+pub mod routes;
+
+pub use route::Route;
+pub use routes::Routes;
