--- src/routing/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2022 VMware, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod router;
